Move lab1 to directory
